+++ app/src/main/res/layout/activity_menu.xml
@@ -1,0 +1,60 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".MenuActivity"
+    android:background="@drawable/menu_background_image">
+
+
+<!--    <TextView-->
+<!--        android:id="@+id/txtgameTitle"-->
+<!--        android:layout_width="match_parent"-->
+<!--        android:layout_height="wrap_content"-->
+<!--        android:layout_marginTop="100dp"-->
+<!--        android:gravity="center_horizontal"-->
+<!--        android:text="@string/app_name"-->
+<!--        android:textSize="50sp"-->
+<!--        app:layout_constraintEnd_toEndOf="parent"-->
+<!--        app:layout_constraintStart_toStartOf="parent"-->
+<!--        app:layout_constraintTop_toTopOf="parent" />-->
+
+    <Button
+        android:id="@+id/btnplay"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/play"
+        android:textSize="50sp"
+        app:cornerRadius="20dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <!--    <Button-->
+<!--        android:id="@+id/btnother"-->
+<!--        android:layout_width="wrap_content"-->
+<!--        android:layout_height="wrap_content"-->
+<!--        android:layout_marginTop="50dp"-->
+<!--        android:text="@string/Other"-->
+<!--        android:textSize="40sp"-->
+<!--        app:cornerRadius="20dp"-->
+<!--        app:layout_constraintEnd_toEndOf="@+id/btnhighscore"-->
+<!--        app:layout_constraintStart_toStartOf="@+id/btnhighscore"-->
+<!--        app:layout_constraintTop_toBottomOf="@+id/btnhighscore" />-->
+
+<!--    <Button-->
+<!--        android:id="@+id/btnhighscore"-->
+<!--        android:layout_width="wrap_content"-->
+<!--        android:layout_height="wrap_content"-->
+<!--        android:layout_marginTop="80dp"-->
+<!--        android:text="@string/HighScore"-->
+<!--        android:textSize="40sp"-->
+<!--        app:cornerRadius="20dp"-->
+<!--        app:layout_constraintEnd_toEndOf="@+id/btnplay"-->
+<!--        app:layout_constraintHorizontal_bias="0.383"-->
+<!--        app:layout_constraintStart_toStartOf="@+id/btnplay"-->
+<!--        app:layout_constraintTop_toBottomOf="@+id/btnplay" />-->
+
+</androidx.constraintlayout.widget.ConstraintLayout>
